refactor(gemini): move native image transport behind compatible facade
